# BAAS-23656: Remove limiter nil check in waitOneTick method