https://jira.mongodb.org/browse/BAAS-23656

Benchmark results:

current
```
goja (BAAS-23656_removeSelfLimiterNilCheck)$ go test -bench=BenchmarkEmptyLoop$ -run=$ -benchtime 1000000x
--
goos: darwin
goarch: arm64
pkg: github.com/dop251/goja
BenchmarkEmptyLoop-10            1000000              6143 ns/op
PASS
ok      github.com/dop251/goja  6.537s
```

new
```
goja (BAAS-23656_removeSelfLimiterNilCheck)$ go test -bench=BenchmarkEmptyLoop$ -run=$ -benchtime 1000000x
--
goos: darwin
goarch: arm64
pkg: github.com/dop251/goja
BenchmarkEmptyLoop-10            1000000              6047 ns/op
PASS
ok      github.com/dop251/goja  6.237s
```

Roughly a 1.5% improvement


Note: there are still a few additional goja tests that I'll need to update to ensure limiter isn't nil, but wanted to gather opinions on this change prior to doing that work.



Edit:
Also inlining the waitOneTick function like Andy suggested gave us a 8.5% improvement

```
goja (BAAS-23656_removeSelfLimiterNilCheck)$ go test -bench=BenchmarkEmptyLoop$ -run=$ -benchtime 1000000x
goos: darwin
goarch: arm64
pkg: github.com/dop251/goja
BenchmarkEmptyLoop-10            1000000              5623 ns/op
PASS
ok      github.com/dop251/goja  6.443s
```